+++ astropy/time/core.py
@@ -172,6 +172,67 @@ class TimeInfo(MixinInfo):
 
         return out
 
+    def new_like(self, cols, length, metadata_conflicts='warn', name=None):
+        """
+        Return a new Time instance which is consistent with the input Time objects
+        ``cols`` and has ``length`` rows.
+
+        This is intended for creating an empty Time instance whose elements can
+        be set in-place for table operations like join or vstack.  It checks
+        that the input locations and attributes are consistent.  This is used
+        when a Time object is used as a mixin column in an astropy Table.
+
+        Parameters
+        ----------
+        cols : list
+            List of input columns (Time objects)
+        length : int
+            Length of the output column object
+        metadata_conflicts : str ('warn'|'error'|'silent')
+            How to handle metadata conflicts
+        name : str
+            Output column name
+
+        Returns
+        -------
+        col : Time (or subclass)
+            Empty instance of this class consistent with ``cols``
+
+        """
+        # Get merged info attributes like shape, dtype, format, description, etc.
+        attrs = self.merge_cols_attributes(cols, metadata_conflicts, name,
+                                           ('meta', 'description'))
+        attrs.pop('dtype')  # Not relevant for Time
+        col0 = cols[0]
+
+        # Check that location is consistent for all Time objects
+        for col in cols[1:]:
+            # This is the method used by __setitem__ to ensure that the right side
+            # has a consistent location (and coerce data if necessary, but that does
+            # not happen in this case since `col` is already a Time object).  If this
+            # passes then any subsequent table operations via setitem will work.
+            try:
+                col0._make_value_equivalent(slice(None), col)
+            except ValueError:
+                raise ValueError('input columns have inconsistent locations')
+
+        # Make a new Time object with the desired shape and attributes
+        shape = (length,) + attrs.pop('shape')
+        jd2000 = 2451544.5  # Arbitrary JD value J2000.0 that will work with ERFA
+        jd1 = np.full(shape, jd2000, dtype='f8')
+        jd2 = np.zeros(shape, dtype='f8')
+        tm_attrs = {attr: getattr(col0, attr)
+                    for attr in ('scale', 'location',
+                                 'precision', 'in_subfmt', 'out_subfmt')}
+        out = self._parent_cls(jd1, jd2, format='jd', **tm_attrs)
+        out.format = col0.format
+
+        # Set remaining info attributes
+        for attr, value in attrs.items():
+            setattr(out.info, attr, value)
+
+        return out
+
 
 class TimeDeltaInfo(TimeInfo):
     _represent_as_dict_extra_attrs = ('format', 'scale')